+++ CHANGELOG.md
@@ -103,6 +103,7 @@
 - domain parameter is not used externally (#319)
 - pminfo.lua fetches and saves epoch but never uses it #320
 - pminfo.lua - no need to use scroll unless doc count > 10000 (#321)
+- pminfo.lua - scroll/scan ignores sort order (#322)
 
 ## CHANGES in 0.9b: